assigment-2/pyproject.toml
[project]
name = "assigment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.10",
    "python-decouple>=3.8",
]
